packages/amprealize/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amprealize"
version = "0.1.0"
description = "Infrastructure-as-Code orchestration with blueprint-driven container management and compliance tracking"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["infrastructure", "containers", "podman", "orchestration", "compliance", "devops", "iac"]
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Topic :: System :: Systems Administration",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
authors = [
  { name = "GuideAI Team", email = "team@guideai.dev" }
]
dependencies = [
  "pydantic>=2.5,<3.0",
  "pyyaml>=6.0,<7.0",
]

[project.urls]
Documentation = "https://github.com/Nas4146/amprealize"
Source = "https://github.com/Nas4146/amprealize"
Issues = "https://github.com/Nas4146/amprealize/issues"

[project.optional-dependencies]
dev = [
  "pytest>=7.4,<9.0",
  "pytest-cov>=4.1,<6.0",
  "pytest-asyncio>=0.21,<1.0",
  "black>=23.0,<25.0",
  "isort>=5.12,<6.0",
  "mypy>=1.5,<2.0",
]
cli = [
  "typer>=0.9,<1.0",
  "rich>=13.0,<14.0",
]
fastapi = [
  "fastapi>=0.110,<1.0",
  "uvicorn>=0.30,<1.0",
]
all = [
  "amprealize[cli,fastapi]",
]

[project.scripts]
amprealize = "amprealize.cli:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
amprealize = ["py.typed", "blueprints/*.yaml", "blueprints/*.yml"]

[tool.pytest.ini_options]
minversion = "7.4"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=amprealize",
    "--cov-report=term-missing",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
